Split import-export into new file
